Renamed party role seed to just role.

+++ tests/KeeperData.Infrastructure.Tests.Unit/Services/MongoDataSeederTests.cs
@@ -9,15 +9,8 @@ using Microsoft.Extensions.Logging;
 using Microsoft.Extensions.Options;
 using MongoDB.Driver;
 using Moq;
-using System;
-using System.Collections.Generic;
-using System.IO;
-using System.Linq;
 using System.Reflection;
 using System.Text.Json;
-using System.Threading;
-using System.Threading.Tasks;
-using Xunit;
 
 namespace KeeperData.Infrastructure.Tests.Unit.Services;
 
@@ -29,7 +22,7 @@ public class MongoDataSeederTests : IDisposable
     private readonly Mock<IMongoDatabase> _mockDatabase;
     private readonly Mock<IMongoCollection<CountryListDocument>> _mockCountryCollection;
     private readonly Mock<IMongoCollection<SpeciesListDocument>> _mockSpeciesCollection;
-    private readonly Mock<IMongoCollection<PartyRoleListDocument>> _mockPartyRoleCollection;
+    private readonly Mock<IMongoCollection<RoleListDocument>> _mockRoleCollection;
     private readonly Mock<IMongoCollection<PremisesTypeListDocument>> _mockPremisesTypeCollection;
     private readonly Mock<IMongoCollection<PremisesActivityTypeListDocument>> _mockPremisesActivityTypeCollection;
     private readonly Mock<IOptions<MongoConfig>> _mockConfig;
@@ -44,7 +37,7 @@ public class MongoDataSeederTests : IDisposable
         _mockDatabase = new Mock<IMongoDatabase>();
         _mockCountryCollection = new Mock<IMongoCollection<CountryListDocument>>();
         _mockSpeciesCollection = new Mock<IMongoCollection<SpeciesListDocument>>();
-        _mockPartyRoleCollection = new Mock<IMongoCollection<PartyRoleListDocument>>();
+        _mockRoleCollection = new Mock<IMongoCollection<RoleListDocument>>();
         _mockPremisesTypeCollection = new Mock<IMongoCollection<PremisesTypeListDocument>>();
         _mockPremisesActivityTypeCollection = new Mock<IMongoCollection<PremisesActivityTypeListDocument>>();
         _mockConfig = new Mock<IOptions<MongoConfig>>();
@@ -58,7 +51,7 @@ public class MongoDataSeederTests : IDisposable
 
         _mockDatabase.Setup(d => d.GetCollection<CountryListDocument>("refCountries", null)).Returns(_mockCountryCollection.Object);
         _mockDatabase.Setup(d => d.GetCollection<SpeciesListDocument>("refSpecies", null)).Returns(_mockSpeciesCollection.Object);
-        _mockDatabase.Setup(d => d.GetCollection<PartyRoleListDocument>("refPartyRoles", null)).Returns(_mockPartyRoleCollection.Object);
+        _mockDatabase.Setup(d => d.GetCollection<RoleListDocument>("refRoles", null)).Returns(_mockRoleCollection.Object);
         _mockDatabase.Setup(d => d.GetCollection<PremisesTypeListDocument>("refPremisesTypes", null)).Returns(_mockPremisesTypeCollection.Object);
         _mockDatabase.Setup(d => d.GetCollection<PremisesActivityTypeListDocument>("refPremisesActivityTypes", null)).Returns(_mockPremisesActivityTypeCollection.Object);
         _mockClient.Setup(c => c.GetDatabase("TestDb", null)).Returns(_mockDatabase.Object);
@@ -69,9 +62,18 @@ public class MongoDataSeederTests : IDisposable
 
     public void Dispose()
     {
-        if (Directory.Exists(_testDirectory))
+        Dispose(true);
+        GC.SuppressFinalize(this);
+    }
+
+    protected virtual void Dispose(bool disposing)
+    {
+        if (disposing)
         {
-            Directory.Delete(_testDirectory, true);
+            if (Directory.Exists(_testDirectory))
+            {
+                Directory.Delete(_testDirectory, true);
+            }
         }
     }
 
@@ -88,7 +90,7 @@ public class MongoDataSeederTests : IDisposable
         return filePath;
     }
 
-    private CountryDocument CreateTestCountry(string code, string name)
+    private static CountryDocument CreateTestCountry(string code, string name)
     {
         return new CountryDocument
         {
@@ -108,7 +110,7 @@ public class MongoDataSeederTests : IDisposable
             LastModifiedDate = null
         };
     }
-    private SpeciesDocument CreateTestSpecies(string code, string name)
+    private static SpeciesDocument CreateTestSpecies(string code, string name)
     {
         return new SpeciesDocument
         {
@@ -126,9 +128,9 @@ public class MongoDataSeederTests : IDisposable
         };
     }
 
-    private PartyRoleDocument CreateTestPartyRole(string code, string name)
+    private static RoleDocument CreateTestRole(string code, string name)
     {
-        return new PartyRoleDocument
+        return new RoleDocument
         {
             IdentifierId = Guid.NewGuid().ToString(),
             Code = code,
@@ -143,7 +145,7 @@ public class MongoDataSeederTests : IDisposable
             LastModifiedDate = null
         };
     }
-    private PremisesTypeDocument CreateTestPremisesType(string code, string name)
+    private static PremisesTypeDocument CreateTestPremisesType(string code, string name)
     {
         return new PremisesTypeDocument
         {
@@ -157,7 +159,7 @@ public class MongoDataSeederTests : IDisposable
         };
     }
 
-    private PremisesActivityTypeDocument CreateTestPremisesActivityType(string code, string name)
+    private static PremisesActivityTypeDocument CreateTestPremisesActivityType(string code, string name)
     {
         return new PremisesActivityTypeDocument
         {
@@ -178,13 +180,13 @@ public class MongoDataSeederTests : IDisposable
         await seeder.StartAsync(CancellationToken.None);
         _mockLogger.VerifyLog(LogLevel.Information, "Seed file 'countries.json' not found", Times.Once());
         _mockLogger.VerifyLog(LogLevel.Information, "Seed file 'species.json' not found", Times.Once());
-        _mockLogger.VerifyLog(LogLevel.Information, "Seed file 'partyroles.json' not found", Times.Once());
+        _mockLogger.VerifyLog(LogLevel.Information, "Seed file 'roles.json' not found", Times.Once());
         _mockLogger.VerifyLog(LogLevel.Information, "Seed file 'premisestypes.json' not found", Times.Once());
         _mockLogger.VerifyLog(LogLevel.Information, "Seed file 'premisesactivitytypes.json' not found", Times.Once());
 
         _mockCountryCollection.VerifyNoOtherCalls();
         _mockSpeciesCollection.VerifyNoOtherCalls();
-        _mockPartyRoleCollection.VerifyNoOtherCalls();
+        _mockRoleCollection.VerifyNoOtherCalls();
         _mockPremisesTypeCollection.VerifyNoOtherCalls();
         _mockPremisesActivityTypeCollection.VerifyNoOtherCalls();
     }
@@ -194,7 +196,7 @@ public class MongoDataSeederTests : IDisposable
     {
         var seeder = CreateSeeder();
 
-        CreateJsonFile("countries.json", new List<CountryDocument> { CreateTestCountry("GB", "UK") });
+        CreateJsonFile("countries.json", [CreateTestCountry("GB", "UK")]);
 
         await seeder.StartAsync(CancellationToken.None);
 
@@ -208,15 +210,15 @@ public class MongoDataSeederTests : IDisposable
     {
         var seeder = CreateSeeder();
 
-        CreateJsonFile("countries.json", new List<CountryDocument> { CreateTestCountry("US", "USA") });
-        CreateJsonFile("species.json", new List<SpeciesDocument> { CreateTestSpecies("CTT", "Cattle") });
-        CreateJsonFile("partyroles.json", new List<PartyRoleDocument> { CreateTestPartyRole("KEEPER", "Livestock Keeper") });
-        CreateJsonFile("premisestypes.json", new List<PremisesTypeDocument> { CreateTestPremisesType("AH", "Agricultural Holding") });
-        CreateJsonFile("premisesactivitytypes.json", new List<PremisesActivityTypeDocument> { CreateTestPremisesActivityType("AFU", "Approved Finishing Unit") });
+        CreateJsonFile("countries.json", [CreateTestCountry("US", "USA")]);
+        CreateJsonFile("species.json", [CreateTestSpecies("CTT", "Cattle")]);
+        CreateJsonFile("roles.json", [CreateTestRole("KEEPER", "Livestock Keeper")]);
+        CreateJsonFile("premisestypes.json", [CreateTestPremisesType("AH", "Agricultural Holding")]);
+        CreateJsonFile("premisesactivitytypes.json", [CreateTestPremisesActivityType("AFU", "Approved Finishing Unit")]);
 
         CountryListDocument? capturedCountryDoc = null;
         SpeciesListDocument? capturedSpeciesDoc = null;
-        PartyRoleListDocument? capturedPartyRoleDoc = null;
+        RoleListDocument? capturedRoleDoc = null;
         PremisesTypeListDocument? capturedPremisesTypeDoc = null;
         PremisesActivityTypeListDocument? capturedPremisesActivityTypeDoc = null;
 
@@ -228,8 +230,8 @@ public class MongoDataSeederTests : IDisposable
                            .Callback<FilterDefinition<SpeciesListDocument>, SpeciesListDocument, ReplaceOptions, CancellationToken>((_, doc, _, _) => capturedSpeciesDoc = doc)
                            .Returns(Task.FromResult(Mock.Of<ReplaceOneResult>()));
 
-        _mockPartyRoleCollection.Setup(x => x.ReplaceOneAsync(It.IsAny<FilterDefinition<PartyRoleListDocument>>(), It.IsAny<PartyRoleListDocument>(), It.IsAny<ReplaceOptions>(), It.IsAny<CancellationToken>()))
-                           .Callback<FilterDefinition<PartyRoleListDocument>, PartyRoleListDocument, ReplaceOptions, CancellationToken>((_, doc, _, _) => capturedPartyRoleDoc = doc)
+        _mockRoleCollection.Setup(x => x.ReplaceOneAsync(It.IsAny<FilterDefinition<RoleListDocument>>(), It.IsAny<RoleListDocument>(), It.IsAny<ReplaceOptions>(), It.IsAny<CancellationToken>()))
+                           .Callback<FilterDefinition<RoleListDocument>, RoleListDocument, ReplaceOptions, CancellationToken>((_, doc, _, _) => capturedRoleDoc = doc)
                            .Returns(Task.FromResult(Mock.Of<ReplaceOneResult>()));
 
         _mockPremisesTypeCollection.Setup(x => x.ReplaceOneAsync(It.IsAny<FilterDefinition<PremisesTypeListDocument>>(), It.IsAny<PremisesTypeListDocument>(), It.IsAny<ReplaceOptions>(), It.IsAny<CancellationToken>()))
@@ -244,7 +246,7 @@ public class MongoDataSeederTests : IDisposable
 
         _mockCountryCollection.Verify(x => x.ReplaceOneAsync(It.IsAny<FilterDefinition<CountryListDocument>>(), It.IsAny<CountryListDocument>(), It.IsAny<ReplaceOptions>(), It.IsAny<CancellationToken>()), Times.Once);
         _mockSpeciesCollection.Verify(x => x.ReplaceOneAsync(It.IsAny<FilterDefinition<SpeciesListDocument>>(), It.IsAny<SpeciesListDocument>(), It.IsAny<ReplaceOptions>(), It.IsAny<CancellationToken>()), Times.Once);
-        _mockPartyRoleCollection.Verify(x => x.ReplaceOneAsync(It.IsAny<FilterDefinition<PartyRoleListDocument>>(), It.IsAny<PartyRoleListDocument>(), It.IsAny<ReplaceOptions>(), It.IsAny<CancellationToken>()), Times.Once);
+        _mockRoleCollection.Verify(x => x.ReplaceOneAsync(It.IsAny<FilterDefinition<RoleListDocument>>(), It.IsAny<RoleListDocument>(), It.IsAny<ReplaceOptions>(), It.IsAny<CancellationToken>()), Times.Once);
         _mockPremisesTypeCollection.Verify(x => x.ReplaceOneAsync(It.IsAny<FilterDefinition<PremisesTypeListDocument>>(), It.IsAny<PremisesTypeListDocument>(), It.IsAny<ReplaceOptions>(), It.IsAny<CancellationToken>()), Times.Once);
         _mockPremisesActivityTypeCollection.Verify(x => x.ReplaceOneAsync(It.IsAny<FilterDefinition<PremisesActivityTypeListDocument>>(), It.IsAny<PremisesActivityTypeListDocument>(), It.IsAny<ReplaceOptions>(), It.IsAny<CancellationToken>()), Times.Once);
 
@@ -254,8 +256,8 @@ public class MongoDataSeederTests : IDisposable
         capturedSpeciesDoc.Should().NotBeNull();
         capturedSpeciesDoc!.Species.Should().Contain(s => s.Code == "CTT");
 
-        capturedPartyRoleDoc.Should().NotBeNull();
-        capturedPartyRoleDoc!.PartyRoles.Should().Contain(pr => pr.Code == "KEEPER");
+        capturedRoleDoc.Should().NotBeNull();
+        capturedRoleDoc!.Roles.Should().Contain(pr => pr.Code == "KEEPER");
 
         capturedPremisesTypeDoc.Should().NotBeNull();
         capturedPremisesTypeDoc?.PremisesTypes.Should().Contain(pt => pt.Code == "AH");
@@ -269,10 +271,10 @@ public class MongoDataSeederTests : IDisposable
     {
         var seeder = CreateSeeder();
 
-        CreateJsonFile("countries.json", new List<CountryDocument> { CreateTestCountry("US", "USA") });
+        CreateJsonFile("countries.json", [CreateTestCountry("US", "USA")]);
 
         _mockCountryCollection.Setup(x => x.ReplaceOneAsync(It.IsAny<FilterDefinition<CountryListDocument>>(), It.IsAny<CountryListDocument>(), It.IsAny<ReplaceOptions>(), It.IsAny<CancellationToken>()))
-                           .ThrowsAsync(new MongoException("Database connection failed"));
+            .ThrowsAsync(new MongoException("Database connection failed"));
 
         await seeder.StartAsync(CancellationToken.None);
